Integrate secure ION buffer into CDMI service to implement SDP
When creating a media engine session, a socket connection is established with
the OpenCDM browser plugin using the socket server helper class. This
connection is used, at every decryption operation, to communicate the secure
(ION) file description from the plugin to the CDMI service.

The secure file descriptor is passed as a new parameter to the
CMediaEngineSession::Decrypt() and CMediaKeySession::Decrypt() API to reach
the DRM-specific implementation.

Multi-session is supported by creating a dedicated socket connection for each
session. For that purpose, an unique socket channel ID is shared with the
OpenCDM browser plugin and it is appended to the UNIX socket address to make
it also unique.

This integration has been tested with the OpenCDM SDP Prototype implemented
at the OpenCDM browser plugin layer. Refer to that layer for more details.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>